+++ lib/widgets/customizable_multiselect_widget.dart
@@ -1,6 +1,7 @@
 import 'package:customizable_multiselect_field/models/customizable_multiselect_dialog_options.dart';
 import 'package:customizable_multiselect_field/models/customizable_multiselect_widget_options.dart';
 import 'package:customizable_multiselect_field/models/data_source.dart';
+import 'package:customizable_multiselect_field/utils/extensions/extended-iterable.dart';
 import 'package:customizable_multiselect_field/widgets/customizable_multiselect_dialog.dart';
 import 'package:flutter/material.dart';
 
@@ -9,15 +10,19 @@ class CustomizableMultiselectWidget extends StatefulWidget {
   const CustomizableMultiselectWidget({
     Key key,
     @required this.dataSourceList,
-    this.onChanged,
     @required this.customizableMultiselectDialogOptions,
     @required this.customizableMultiselectWidgetOptions,
+    @required this.decoration,
+    @required this.onChanged,
+    @required this.value
   }) : super(key: key);
 
   final List<DataSource> dataSourceList;
-  final ValueChanged<List<List<dynamic>>> onChanged;
   final CustomizableMultiselectDialogOptions customizableMultiselectDialogOptions;
   final CustomizableMultiselectWidgetOptions customizableMultiselectWidgetOptions;
+  final InputDecoration decoration;
+  final ValueChanged<List<List<dynamic>>> onChanged;
+  final List<List<dynamic>> value;
 
   @override
   _CustomizableMultiselectWidgetState createState() => _CustomizableMultiselectWidgetState();
@@ -26,7 +31,7 @@ class CustomizableMultiselectWidget extends StatefulWidget {
 
 class _CustomizableMultiselectWidgetState extends State<CustomizableMultiselectWidget> {
 
-  List<Widget> _buildListChip(DataSource dataSource) => dataSource.valueList
+  List<Widget> _buildListChip(DataSource dataSource, int index) => widget.value[index]
     .map((value) => dataSource.dataList.singleWhere((data) => data[dataSource.options.valueKey] == value,orElse: () => null))
     .map((value) => (value != null)
       ? Chip(
@@ -44,6 +49,22 @@ class _CustomizableMultiselectWidgetState extends State<CustomizableMultiselectW
     return widget.dataSourceList.every((DataSource element) => element.valueList.isEmpty);
   }
 
+
+  InputDecoration _getEffectiveDecoration() {
+    final ThemeData themeData = Theme.of(context);
+    final InputDecoration effectiveDecoration = widget.decoration
+      .applyDefaults(themeData.inputDecorationTheme)
+      .copyWith(
+        enabled: widget.customizableMultiselectWidgetOptions.enable,
+        hintMaxLines: widget.decoration?.hintMaxLines,
+      );
+    return effectiveDecoration.copyWith(
+      errorText: effectiveDecoration.errorText ?? '',
+      counterStyle: effectiveDecoration.errorStyle
+        ?? themeData.textTheme.caption.copyWith(color: themeData.errorColor),
+    );
+  }
+
   @override
   Widget build(BuildContext context) {
     return InkWell(
@@ -53,6 +74,7 @@ class _CustomizableMultiselectWidgetState extends State<CustomizableMultiselectW
             context: context,
             builder: (BuildContext context) {
               return CustomizableMultiselectDialog(
+                selectedValues: widget.value,
                 dataSourceList: widget.dataSourceList,
                 customizableMultiselectDialogOptions: widget.customizableMultiselectDialogOptions,
               );
@@ -64,12 +86,12 @@ class _CustomizableMultiselectWidgetState extends State<CustomizableMultiselectW
         }
       },
       child: InputDecorator(
-        decoration: widget.customizableMultiselectWidgetOptions.decoration,
+        decoration: _getEffectiveDecoration(),
         child: (!_checkIfAllDataSourceIsEmpty())
             ? Column(
                 children: <Widget>[
-                  ...widget.dataSourceList.map(
-                    (DataSource dataSource) => (dataSource.valueList.isNotEmpty)
+                  ...widget.dataSourceList.mapIndex(
+                    (DataSource dataSource, int index) => (widget.value[index].isNotEmpty)
                     ? Padding(
                       padding: const EdgeInsets.all(8.0),
                       child: Column(
@@ -92,7 +114,7 @@ class _CustomizableMultiselectWidgetState extends State<CustomizableMultiselectW
                                   child: Wrap(
                                   spacing: 8.0,
                                   children:
-                                      _buildListChip(dataSource),
+                                      _buildListChip(dataSource, index),
                                 ))
                               ],
                             ),